[REFACTOR] #267 게시글 삭제 mutation 이동 및 삭제 시 새로고침 수정
